requirements.txt
aniso8601==9.0.1
blinker==1.6.2
click==8.1.3
Flask==2.3.2
Flask-RESTful==0.3.10
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pika==1.3.2
python-dotenv==1.0.0
blessed==1.20.0
pytz==2023.3
six==1.16.0
Werkzeug==2.3.4
